# spec/openapi/apx.yaml
openapi: 3.1.0
info:
  title: APX — APDS Parking eXtensions
  version: 0.1.0
  summary: An open companion standard to APDS 4.1
  description: |-
    **APX (APDS Parking eXtensions)** is an additive companion standard to the
    [APDS 4.1 API](https://github.com/parkingdata/spec) (Alliance for Parking
    Data Standards, published as ISO/TS 5206-1).

    APX follows an **APDS-first prime directive**: wherever APDS 4.1 already
    defines a route, schema, or convention, APX uses it verbatim (by `$ref`
    into the vendored, unmodified `apds-api-4.1.yaml`). APX only adds what
    APDS deliberately leaves out:

    - **Data profile** — Full/Change update semantics, tombstones, and a
      cursor-based change feed as additive parameters on APDS's native routes
    - **Delivery fabric** — a superset-compatible profile of APDS `/webhooks`
      (HMAC signing, normative retries, delivery ledger, SSE)
    - **Control** — command plane for gates/PARCS/devices with idempotency,
      lifecycle, and immutable audit
    - **Alerts** — operational alerting with open taxonomies via APDS
      UserDefinedCodeLists
    - **Discovery** — credential-scoped capability documents
    - **Accounts & payments, LPR, reservations, permits, tolling** — profiles
      and additive resources over APDS entities

    Extension identity follows the official APDS extension convention:
    `apds-ext:apx:<class>@<version>` (Use Case Document §C.2.5), with
    extensionName `apx`.
  license:
    name: MIT
...
servers:
...
paths:
...
  /apx/v1/commands:
    post:
      operationId: post-apx-v1-commands
      tags: [Control]
      summary: Execute a command (idempotent, perishable, audited)
      description: |-
        Requests an operational action (vendGate, lostTicket, pushRate,
        applyValidation, …). Idempotency-Key REQUIRED. Returns 202 with the
        Command; poll GET /apx/v1/commands/{id} or subscribe to
        apx.control.command.status.v1. Requires scope apx.control:execute
        AND a place grant covering the target.
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Command"
      responses:
        "202":
          description: Command accepted for asynchronous execution.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Command"
        "400":
          description: Missing Idempotency-Key, unknown commandType, or missing target.
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "403":
          description: Insufficient scope or place grant.
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "409":
          description: Idempotency-Key replayed with a different body.
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
  /apx/v1/commands/{id}:
    get:
      operationId: get-apx-v1-commands-id
      tags: [Control]
      summary: Read a command (lifecycle + audit trail)
      description: Returns the Command including its immutable statusHistory.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: The command.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Command" }
        "404":
          description: No such command.
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
  /apx/v1/commands/{id}/cancel:
    post:
      operationId: post-apx-v1-commands-id-cancel
      tags: [Control]
      summary: Cancel a command before dispatch
      description: Allowed in received/accepted states; afterwards 409 command-not-cancellable.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Cancelled command.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Command" }
        "409":
          description: Not cancellable from the current state.
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
...
  /apx/v1/devices/{id}:
    get:
      operationId: get-apx-v1-devices-id
      tags: [Control]
      summary: Read one device's status overlay
      description: DeviceStatus keyed by the SupplementalEquipment id.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Device status.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DeviceStatus" }
        "404":
          description: No such device.
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
  /apx/v1/lanes/{id}/current:
    get:
      operationId: get-apx-v1-lanes-id-current
      tags: [Control]
      summary: Lane inquiry — current transaction context (screen-pop)
      description: |-
        The ticket/credential currently in the machine at this lane, amount
        due, applied validations, latest LPR read, and monthly-credential
        context including the last 10 activity events (2018 requirement ①).
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
          description: VehicularAccess (lane) UUID.
      responses:
        "200":
          description: Current lane status.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LaneStatus" }
        "404":
          description: No such lane or no live state.
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
  /apx/v1/validations/providers:
    get:
      operationId: get-apx-v1-validations-providers
      tags: [Control]
      summary: List validation providers for a place
      description: The venues allowed to validate tickets at the place (2018 requirement ⑤).
      parameters:
        - name: place
          in: query
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Validation providers.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/ValidationProvider" }

  # ===== Alerts (conformance class apx-alerts) ===== #
...
security:
  - apxOAuth: []
...
components:
...
  securitySchemes:
    apxOAuth:
      type: oauth2
      description: |-
        OAuth 2.0 client-credentials (aligned with APDS 4.1's `oAuth` scheme).
        APX scopes follow `apx.<domain>:<verb>`. Grants are carried in the
        normative token claims `apx_org` (Organisation reference) and
        `apx_places` (HierarchyElement UUID subtree grants; absent = all).
      flows:
        clientCredentials:
          tokenUrl: https://auth.example.com/oauth2/token
          scopes:
            apx.data:read: Read parking data via APDS-native routes
            apx.data:write: Create/update parking data via APDS-native routes
            apx.control:read: Read commands, lanes, and device status
            apx.control:execute: Execute commands (vend gate, lost ticket, push rate, apply validation…)
            apx.alerts:read: Read alerts
            apx.alerts:write: Raise, acknowledge, and resolve alerts
            apx.subscriptions:manage: Create and manage event subscriptions
            apx.accounts:read: Look up accounts and balances
            apx.payments:write: Take payments and post accounting write-backs
            apx.lpr:read: LPR plate/ticket cross-lookup
            apx.reservations:manage: Quote, book, amend, cancel reservations
            apx.permits:manage: Issue and manage permits
            apx.tolling:manage: Create and manage toll transactions
  schemas:
...
    Problem:
      $ref: "./domains/common/schemas.yaml#/Problem"
...
    Command:
      $ref: "./domains/control/schemas.yaml#/Command"
    DeviceStatus:
      $ref: "./domains/control/schemas.yaml#/DeviceStatus"
    LaneStatus:
      $ref: "./domains/control/schemas.yaml#/LaneStatus"
    ValidationProvider:
      $ref: "./domains/control/schemas.yaml#/ValidationProvider"
...